+++ cmd/dep/ensure.go
@@ -163,6 +163,7 @@ func (cmd *ensureCommand) Run(ctx *dep.Ctx, args []string) error {
 	if err != nil {
 		return err
 	}
+	ctx.Registry = gps.NewFileRegistry(p.AbsRoot)
 
 	sm, err := ctx.SourceManager()
 	if err != nil {